+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Unit tests for `pyodide` runner (#976)
 - Dispatch event when project identifier changes, e.g. after project is remixed (#2830)
 - Remove broken `format` script (#991)
+- Invalidate cached project when project is remixed (#1003)
 
 ## [0.22.2] - 2024-03-18